moved registration form to components folder
