+++ src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import './App.css';
 
 import Login from './components/Login';
-import RegisterForm from './RegisterForm'
+import RegisterForm from './components/RegisterForm';
 
 function App() {
   return (
